Move JsonUtils->gson, JarUtils->io, Properties->javafx
