+++ HMCL/src/main/java/org/jackhuang/hmcl/game/HMCLModpackManager.java
@@ -19,9 +19,9 @@ package org.jackhuang.hmcl.game;
 
 import com.google.gson.JsonParseException;
 import org.jackhuang.hmcl.mod.Modpack;
-import org.jackhuang.hmcl.util.JsonUtils;
 import org.jackhuang.hmcl.util.Lang;
 import org.jackhuang.hmcl.util.StringUtils;
+import org.jackhuang.hmcl.util.gson.JsonUtils;
 import org.jackhuang.hmcl.util.io.CompressingUtils;
 
 import java.io.File;
