+++ HMCLCore/src/main/java/org/jackhuang/hmcl/mod/CurseManifest.java
@@ -21,7 +21,7 @@ import com.google.gson.JsonParseException;
 import com.google.gson.annotations.SerializedName;
 
 import org.jackhuang.hmcl.util.Immutable;
-import org.jackhuang.hmcl.util.JsonUtils;
+import org.jackhuang.hmcl.util.gson.JsonUtils;
 import org.jackhuang.hmcl.util.io.CompressingUtils;
 
 import java.io.File;
