+++ HMCLCore/src/main/java/org/jackhuang/hmcl/util/gson/JsonUtils.java
@@ -1,7 +1,7 @@
 /*
  * Hello Minecraft! Launcher.
  * Copyright (C) 2018  huangyuhui <huanghongxun2008@126.com>
- * 
+ *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
@@ -15,24 +15,23 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see {http://www.gnu.org/licenses/}.
  */
-package org.jackhuang.hmcl.util;
+package org.jackhuang.hmcl.util.gson;
 
-import javafx.beans.property.Property;
+import org.jackhuang.hmcl.util.Constants;
 
-import java.util.concurrent.atomic.AtomicReference;
+import com.google.gson.JsonParseException;
 
 /**
- *
- * @author huangyuhui
+ * @author yushijinhun
  */
-public final class Properties {
+public final class JsonUtils {
 
-    private Properties() {
-    }
+    private JsonUtils() {}
 
-    public static <T> void updateAsync(Property<? super T> property, T newValue, AtomicReference<T> update) {
-        if (update.getAndSet(newValue) == null)
-            Constants.UI_THREAD_SCHEDULER.accept(() ->
-                    property.setValue(update.getAndSet(null)));
+    public static <T> T fromNonNullJson(String json, Class<T> classOfT) throws JsonParseException {
+        T parsed = Constants.GSON.fromJson(json, classOfT);
+        if (parsed == null)
+            throw new JsonParseException("Json object cannot be null.");
+        return parsed;
     }
 }
