+++ HMCLCore/src/main/java/org/jackhuang/hmcl/util/io/JarUtils.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see {http://www.gnu.org/licenses/}.
  */
-package org.jackhuang.hmcl.util;
+package org.jackhuang.hmcl.util.io;
 
 import java.io.IOException;
 import java.net.URISyntaxException;
